Move merge_nodes to avoid circular import
