+++ fil_finder/length.py
@@ -1,8 +1,6 @@
 # Licensed under an MIT open source license - see LICENSE
 
 from .utilities import *
-from .pixel_ident import *
-
 
 import numpy as np
 import scipy.ndimage as nd
@@ -790,3 +788,22 @@ def main_length(max_path, edge_list, labelisofil, interpts, branch_lengths,
                 p.clf()
 
     return main_lengths, longpath_arrays
+
+
+def merge_nodes(node, G):
+    '''
+    Combine a node into its neighbors.
+    '''
+
+    neigb = list(G[node])
+
+    if len(neigb) != 2:
+        return G
+
+    new_weight = G[node][neigb[0]]['weight'] + \
+        G[node][neigb[1]]['weight']
+
+    G.remove_node(node)
+    G.add_edge(neigb[0], neigb[1], weight=new_weight)
+
+    return G
